fix: remove codeHostings when deleting publishers
Remove the associated codeHostings when using DELETE on a publisher.

Code hostings are mapped 1-to-1 with a publisher, so they they're part
of the resource.
Not deleting them means we'd have dangling code hosting with no
publisher, invisible to the API, and that can't be associated to a
publisher anymore.